Abstract ModelTextures for use with both blocks and items
