+++ rainbow/src/main/java/org/geysermc/rainbow/mapping/PackContext.java
@@ -3,6 +3,8 @@ package org.geysermc.rainbow.mapping;
 import org.geysermc.rainbow.definition.GeyserMappings;
 import org.geysermc.rainbow.mapping.geometry.GeometryRenderer;
 import org.geysermc.rainbow.mapping.geometry.MappedGeometryCache;
+import org.geysermc.rainbow.mapping.texture.BlockModelTextures;
+import org.geysermc.rainbow.mapping.texture.ItemModelTextures;
 import org.geysermc.rainbow.mapping.texture.ModelTextureCache;
 import org.geysermc.rainbow.pack.PackPaths;
 
@@ -16,7 +18,8 @@ public final class PackContext {
     private final AssetResolver assetResolver;
     private final Optional<GeometryRenderer> geometryRenderer;
     private final boolean reportSuccesses;
-    private final ModelTextureCache textureCache = new ModelTextureCache();
+    private final ModelTextureCache<BlockModelTextures> blockTextureCache = new ModelTextureCache<>();
+    private final ModelTextureCache<ItemModelTextures> itemTextureCache = new ModelTextureCache<>();
     private final MappedGeometryCache geometryCache = new MappedGeometryCache();
 
     public PackContext(GeyserMappings mappings, PackPaths paths, BedrockAssetConsumer assetConsumer, AssetResolver assetResolver,
@@ -53,8 +56,12 @@ public final class PackContext {
         return reportSuccesses;
     }
 
-    public ModelTextureCache textureCache() {
-        return textureCache;
+    public ModelTextureCache<BlockModelTextures> blockTextureCache() {
+        return blockTextureCache;
+    }
+
+    public ModelTextureCache<ItemModelTextures> itemTextureCache() {
+        return itemTextureCache;
     }
 
     public MappedGeometryCache geometryCache() {
@@ -62,6 +69,6 @@ public final class PackContext {
     }
 
     public AssetCacheStats cacheStats() {
-        return new AssetCacheStats(geometryCache.stats(), textureCache.stats());
+        return new AssetCacheStats(geometryCache.stats(), itemTextureCache.stats());
     }
 }
