+++ rainbow/src/main/java/org/geysermc/rainbow/mapping/attachable/BedrockAttachableContext.java
@@ -12,7 +12,7 @@ import org.geysermc.rainbow.mapping.PackSerializer;
 import org.geysermc.rainbow.mapping.PackSerializingContext;
 import org.geysermc.rainbow.mapping.geometry.BedrockGeometryContext;
 import org.geysermc.rainbow.mapping.geometry.MappedGeometry;
-import org.geysermc.rainbow.mapping.texture.ModelTextures;
+import org.geysermc.rainbow.mapping.texture.ItemModelTextures;
 import org.geysermc.rainbow.mapping.texture.TextureHolder;
 import org.geysermc.rainbow.pack.PackPaths;
 import org.geysermc.rainbow.pack.attachable.BedrockAttachable;
@@ -40,7 +40,7 @@ public record BedrockAttachableContext(Optional<BedrockAttachable> attachable, O
                 .save(context);
     }
 
-    public static BedrockAttachableContext create(Identifier identifier, ItemStackTemplate stack, BedrockGeometryContext geometryContext, ModelTextures textures, PackContext context) {
+    public static BedrockAttachableContext create(Identifier identifier, ItemStackTemplate stack, BedrockGeometryContext geometryContext, ItemModelTextures textures, PackContext context) {
         // Prefer equippable over animation or geometry attachable, since when an item is equippable, it shows its 2D icon in first and third person (see notes in AnimationMapper)
         Equippable equippable = stack.get(DataComponents.EQUIPPABLE);
         if (equippable != null) {
