+++ rainbow/src/main/java/org/geysermc/rainbow/pack/BedrockItem.java
@@ -5,11 +5,11 @@ import org.geysermc.rainbow.mapping.PackSerializer;
 import org.geysermc.rainbow.mapping.PackSerializingContext;
 import org.geysermc.rainbow.mapping.attachable.BedrockAttachableContext;
 import org.geysermc.rainbow.mapping.geometry.BedrockGeometryContext;
-import org.geysermc.rainbow.mapping.texture.ModelTextures;
+import org.geysermc.rainbow.mapping.texture.ItemModelTextures;
 
 import java.util.concurrent.CompletableFuture;
 
-public record BedrockItem(Identifier identifier, String textureName, ModelTextures textures, BedrockGeometryContext geometryContext,
+public record BedrockItem(Identifier identifier, String textureName, ItemModelTextures textures, BedrockGeometryContext geometryContext,
                           BedrockAttachableContext attachableContext) implements PackSerializer.Serializable {
 
     @Override
